+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.2.0
+ - Changes:
+    - **DEPRECATED**: `double.roundPrecision` is deprecated. Use `double.roundDouble` instead of it
+
+## 2.1.0
+ - Add: 
+    - `string.isNotNullEmpty`
+ - Changes:
+    - Make String nullable
+
 ## 2.0.1
 - Update `README.md`